Move CSV constants to top level module
